requirements.txt
acme==1.12.0
affine==2.4.0
amqp==2.5.2
appdirs==1.4.0
APScheduler==3.2.0
asgiref==3.9.1
attrs==25.3.0
billiard==3.6.3.0
boto3==1.14.14
botocore==1.17.63
celery==4.4.0
certbot==1.12.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
ConfigArgParse==1.2.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
distro==1.5.0
Django==3.0.14
django-appconf==1.0.5
django-codemirror2==0.2
django-colorfield==0.1.15
django-cors-headers==3.4.0
django-filter==2.4.0
django-guardian==2.4.0
django-imagekit==4.0.1
django-redis==4.12.1
django-webpack-loader==0.6.0
djangorestframework==3.13.1
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==5.2.2
docutils==0.15.2
drf-nested-routers==0.11.1
drf-yasg==1.20.0
eventlet==0.32.0 ; sys_platform == "win32"
ffmpeg-python==0.2.0
future==1.0.0
GDAL==3.2.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==22.0.0
idna==2.10
importlib-metadata==8.7.0
inflection==0.5.1
itypes==1.1.0
jinja2==3.1.6
jmespath==0.10.0
josepy==1.2.0
kombu==4.6.7
Markdown==3.3.4
MarkupSafe==3.0.2
morecantile==2.1.4
numexpr==2.10.2
numpy==1.26.2
olefile==0.44
openapi-codec==1.1.7
opentelemetry-api==1.36.0
opentelemetry-distro==0.57b0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-celery==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-psycopg2==0.57b0
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==21.3
parsedatetime==2.6
piexif==1.0.13
pilkit==2.0
Pillow==8.3.2
pip-autoremove==0.9.0
protobuf==6.31.1
psycopg2-binary==2.8.6
pydantic==1.10.8
PyJWT==2.10.1
pyodm==1.5.11
pyOpenSSL==20.0.1
pyparsing==3.2.3
pyRFC3339==1.1
pystac==1.10.1
python-dateutil==2.9.0.post0
pytz==2020.1
rasterio==1.2.9 ; sys_platform == 'linux' or sys_platform == 'darwin'
https://github.com/OpenDroneMap/WebODM/releases/download/v1.9.7/rasterio-1.2.10-cp39-cp39-win_amd64.whl ; sys_platform == "win32"
https://github.com/OpenDroneMap/WebODM/releases/download/v1.9.7/GDAL-3.3.3-cp39-cp39-win_amd64.whl ; sys_platform == "win32"
rcssmin==1.0.6
redis==3.2.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3987==1.3.7
rio-cogeo==2.3.1
rio-color==1.0.4
rio-mucho==1.0.0
rio-tiler==2.1.2
rjsmin==1.0.12
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
Shapely==1.7.0 ; sys_platform == "win32"
s3transfer==0.3.7
scipy==1.11.3
simplejson==3.10.0
six==1.17.0
snuggs==1.4.7
sqlparse==0.5.3
strict-rfc3339==0.7
typing-extensions==4.14.1
tzlocal==1.3
uritemplate==3.0.0
urllib3==1.25.11
vine==1.3.0
webcolors==1.5
wrapt==1.17.2
zipp==3.23.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
